rename core to base in gen-types.
