+++ util/gen-types/Cargo.toml
@@ -12,7 +12,7 @@ repository = "https://github.com/nervosnetwork/ckb"
 ckb-hash = {path = "../hash", version = "= 0.112.0-pre"}
 
 [features]
-default = ["std"]
+default = ["calc-hash", "check-data", "serialized-size", "std"]
 calc-hash = ["blake2b-ref"]
 check-data = []
 capacity = []
@@ -25,5 +25,3 @@ ckb-fixed-hash = { path = "../fixed-hash", version = "= 0.112.0-pre", optional =
 numext-fixed-uint = { version = "0.1", features = ["support_rand", "support_heapsize", "support_serde"], optional = true}
 cfg-if = "1.0"
 blake2b-ref = { version = "0.3.1", default-features = false, optional = true }
-ckb-occupied-capacity = { path = "../occupied-capacity", version = "= 0.112.0-pre" }
-ckb-error = { path = "../../error", version = "= 0.112.0-pre" }
