+++ util/gen-types/src/base/mod.rs
@@ -1,0 +1,8 @@
+mod types;
+pub use types::*;
+
+#[cfg(feature = "calc-hash")]
+mod view;
+
+#[cfg(feature = "calc-hash")]
+pub use view::*;
