+++ util/gen-types/src/base/view/std_env.rs
@@ -1,4 +1,4 @@
-use crate::core::ExtraHashView;
+use crate::base::ExtraHashView;
 
 impl ::std::fmt::Display for ExtraHashView {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
