+++ util/gen-types/src/conversion/blockchain/mod.rs
@@ -3,26 +3,22 @@ mod std_env;
 #[cfg(feature = "std")]
 pub use std_env::*;
 
-use crate::{
-    bytes::Bytes,
-    core::{self, Capacity},
-    generated::packed,
-    prelude::*,
-    vec::Vec,
-};
+use crate::{bytes::Bytes, generated::packed, prelude::*, vec::Vec};
 
-impl Pack<packed::Uint64> for Capacity {
-    fn pack(&self) -> packed::Uint64 {
-        self.as_u64().pack()
-    }
-}
+// use ckb_occupied_capacity::Capacity;
 
-impl<'r> Unpack<core::Capacity> for packed::Uint64Reader<'r> {
-    fn unpack(&self) -> core::Capacity {
-        Capacity::shannons(self.unpack())
-    }
-}
-impl_conversion_for_entity_unpack!(Capacity, Uint64);
+// impl Pack<packed::Uint64> for Capacity {
+//     fn pack(&self) -> packed::Uint64 {
+//         self.as_u64().pack()
+//     }
+// }
+
+// impl<'r> Unpack<Capacity> for packed::Uint64Reader<'r> {
+//     fn unpack(&self) -> Capacity {
+//         Capacity::shannons(self.unpack())
+//     }
+// }
+// impl_conversion_for_entity_unpack!(Capacity, Uint64);
 
 impl Pack<packed::Byte32> for [u8; 32] {
     fn pack(&self) -> packed::Byte32 {
@@ -75,7 +71,7 @@ impl Unpack<Bytes> for packed::Bytes {
     }
 }
 
-impl_conversion_for_vector!(Capacity, Uint64Vec, Uint64VecReader);
+// impl_conversion_for_vector!(Capacity, Uint64Vec, Uint64VecReader);
 impl_conversion_for_vector!(Bytes, BytesVec, BytesVecReader);
 impl_conversion_for_packed_optional_pack!(Byte32, Byte32Opt);
 impl_conversion_for_packed_optional_pack!(CellOutput, CellOutputOpt);
