+++ util/gen-types/src/extension/calc_hash.rs
@@ -1,7 +1,10 @@
-use crate::core::ExtraHashView;
-use crate::generated::packed;
-use crate::prelude::*;
-use crate::util::hash::{blake2b_256, new_blake2b};
+use crate::{
+    base::ExtraHashView,
+    generated::packed,
+    prelude::*,
+    util::hash::{blake2b_256, new_blake2b},
+    vec::Vec,
+};
 
 /*
  * Calculate simple hash for packed bytes wrappers.
@@ -18,7 +21,7 @@ where
     R: Reader<'r>,
 {
     fn calc_hash(&self) -> packed::Byte32 {
-        blake2b_256(self.as_slice()).pack().into()
+        blake2b_256(self.as_slice()).pack()
     }
 }
 
@@ -261,8 +264,8 @@ impl<'r> packed::BlockReader<'r> {
     /// - If there is no extension, extra hash is the same as the uncles hash.
     /// - If there is a extension, then extra hash it the hash of the combination
     /// of uncles hash and the extension hash.
-    pub fn calc_extra_hash(&self) -> crate::core::ExtraHashView {
-        crate::core::ExtraHashView::new(self.calc_uncles_hash(), self.calc_extension_hash())
+    pub fn calc_extra_hash(&self) -> crate::base::ExtraHashView {
+        crate::base::ExtraHashView::new(self.calc_uncles_hash(), self.calc_extension_hash())
     }
 
     /// Calculates transaction hashes for all transactions in the block.
