+++ util/gen-types/src/extension/mod.rs
@@ -1,7 +1,7 @@
 #[cfg(feature = "calc-hash")]
 mod calc_hash;
-#[cfg(feature = "capacity")]
-mod capacity;
+// #[cfg(feature = "capacity")]
+// mod capacity;
 #[cfg(feature = "check-data")]
 mod check_data;
 #[cfg(feature = "serialized-size")]
